editoast: move rolling stock metadata
