+++ editoast/src/modelsv2/light_rolling_stock.rs
@@ -10,6 +10,7 @@ use editoast_derive::ModelV2;
 use editoast_schemas::rolling_stock::EnergySource;
 use editoast_schemas::rolling_stock::Gamma;
 use editoast_schemas::rolling_stock::RollingResistance;
+use editoast_schemas::rolling_stock::RollingStockMetadata;
 use editoast_schemas::rolling_stock::RollingStockSupportedSignalingSystems;
 use serde::Serialize;
 use utoipa::ToSchema;
@@ -21,7 +22,6 @@ use crate::modelsv2::Model;
 use crate::schema::rolling_stock::light_rolling_stock::LightEffortCurves;
 use crate::schema::rolling_stock::light_rolling_stock::LightRollingStock;
 use crate::schema::rolling_stock::light_rolling_stock::LightRollingStockWithLiveriesModel;
-use crate::schema::rolling_stock::RollingStockMetadata;
 use crate::schema::track_section::LoadingGaugeType;
 use crate::views::pagination::Paginate;
 use crate::views::pagination::PaginatedResponse;
